requirements.txt
Django==1.10.4
django-pure-pagination==0.3.0
PyMySQL==0.7.9